# docs/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Operator API",
    "description": "REST API for managing issue types and collections in the Operator system.",
    "contact": {
      "name": "gbqr.us",
      "url": "https://github.com/untra/operator"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.30"
  },
  "paths": {
...
    "/api/v1/collections/{name}": {
      "get": {
        "tags": [
          "Collections"
        ],
        "summary": "Get a single collection by name",
        "operationId": "get_one",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Collection name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/delegators/{name}": {
      "get": {
        "tags": [
          "Delegators"
        ],
        "summary": "Get a single delegator by name",
        "operationId": "get_one",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Delegator name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delegator details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegatorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Delegator not found"
          }
        }
      },
      "put": {
        "tags": [
          "Delegators"
        ],
        "summary": "Update an existing delegator",
        "operationId": "update",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Delegator name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDelegatorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Delegator updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegatorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Delegator not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Delegators"
        ],
        "summary": "Delete a delegator by name",
        "operationId": "delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Delegator name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delegator deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegatorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Delegator not found"
          }
        }
      }
    },
...
    "/api/v1/issuetypes/{key}": {
      "get": {
        "tags": [
          "Issue Types"
        ],
        "summary": "Get a single issue type by key",
        "operationId": "get_one",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Issue type key (e.g., FEAT, FIX)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Issue type details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueTypeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Issue type not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Issue Types"
        ],
        "summary": "Update an existing issue type",
        "operationId": "update",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Issue type key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIssueTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Issue type updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueTypeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Cannot modify builtin type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Issue type not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Issue Types"
        ],
        "summary": "Delete an issue type",
        "operationId": "delete",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Issue type key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Issue type deleted"
          },
          "403": {
            "description": "Cannot delete builtin type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Issue type not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/issuetypes/{key}/steps/{step_name}": {
      "get": {
        "tags": [
          "Steps"
        ],
        "summary": "Get a single step by name",
        "operationId": "get_one",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Issue type key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "step_name",
            "in": "path",
            "description": "Step name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Step details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepResponse"
                }
              }
            }
          },
          "404": {
            "description": "Issue type or step not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Steps"
        ],
        "summary": "Update a step",
        "operationId": "update",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Issue type key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "step_name",
            "in": "path",
            "description": "Step name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Step updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Cannot modify builtin type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Issue type or step not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/model-servers": {
      "get": {
        "tags": [
          "ModelServers"
        ],
        "summary": "List all model servers (user-declared + implicit builtins)",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "List of model servers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelServersResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ModelServers"
        ],
        "summary": "Create a new model server",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModelServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Model server created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelServerResponse"
                }
              }
            }
          },
          "409": {
            "description": "Model server already exists"
          }
        }
      }
    },
    "/api/v1/model-servers/{name}": {
      "get": {
        "tags": [
          "ModelServers"
        ],
        "summary": "Get a single model server by name",
        "operationId": "get_one",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Model server name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model server details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelServerResponse"
                }
              }
            }
          },
          "404": {
            "description": "Model server not found"
          }
        }
      },
      "delete": {
        "tags": [
          "ModelServers"
        ],
        "summary": "Delete a user-declared model server by name",
        "description": "Implicit builtin servers cannot be deleted.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Model server name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model server deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelServerResponse"
                }
              }
            }
          },
          "404": {
            "description": "Model server not found"
          },
          "409": {
            "description": "Cannot delete implicit builtin server"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CollectionResponse": {
        "type": "object",
        "description": "Response for a collection",
        "required": [
          "name",
          "description",
          "types",
          "is_active"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "types": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CreateDelegatorRequest": {
        "type": "object",
        "description": "Request to create a new delegator",
        "required": [
          "name",
          "llm_tool",
          "model"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional display name"
          },
          "launch_config": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DelegatorLaunchConfigDto",
                "description": "Optional launch configuration"
              }
            ]
          },
          "llm_tool": {
            "type": "string",
            "description": "LLM tool name (must match a detected tool)"
          },
          "model": {
            "type": "string",
            "description": "Model alias"
          },
          "model_properties": {
            "type": "object",
            "description": "Arbitrary model properties",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "model_server": {
            "type": [
              "string",
              "null"
            ],
            "description": "Name of a declared `ModelServer`. `None` means use the `llm_tool`'s implicit vendor default."
          },
          "name": {
            "type": "string",
            "description": "Unique name for the delegator"
          }
        }
      },
...
      "CreateModelServerRequest": {
        "type": "object",
        "description": "Request to create a new model server",
        "required": [
          "name",
          "kind"
        ],
        "properties": {
          "api_key_env": {
            "type": [
              "string",
              "null"
            ],
            "description": "Name of an env var providing the API key"
          },
          "base_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Base URL of the inference endpoint"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional display name for UI"
          },
          "extra_env": {
            "type": "object",
            "description": "Additional environment variables",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "kind": {
            "type": "string",
            "description": "Kind: \"ollama\", \"openai-compat\", \"anthropic-api\", \"openai-api\", \"google-api\", \"lmstudio\""
          },
          "name": {
            "type": "string",
            "description": "Unique name for this model server"
          }
        }
      },
...
      "DelegatorResponse": {
        "type": "object",
        "description": "Response for a single delegator",
        "required": [
          "name",
          "llm_tool",
          "model",
          "model_properties"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional display name"
          },
          "launch_config": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DelegatorLaunchConfigDto",
                "description": "Optional launch configuration"
              }
            ]
          },
          "llm_tool": {
            "type": "string",
            "description": "LLM tool name (e.g., \"claude\")"
          },
          "model": {
            "type": "string",
            "description": "Model alias (e.g., \"opus\")"
          },
          "model_properties": {
            "type": "object",
            "description": "Arbitrary model properties",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "model_server": {
            "type": [
              "string",
              "null"
            ],
            "description": "Name of a declared `ModelServer`. `None` means use the `llm_tool`'s implicit vendor default."
          },
          "name": {
            "type": "string",
            "description": "Unique name"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Error response body",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "IssueTypeResponse": {
        "type": "object",
        "description": "Response for a single issue type",
        "required": [
          "key",
          "name",
          "description",
          "mode",
          "glyph",
          "project_required",
          "source",
          "fields",
          "steps"
        ],
        "properties": {
          "color": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldResponse"
            }
          },
          "glyph": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "project_required": {
            "type": "boolean"
          },
          "source": {
            "type": "string"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StepResponse"
            }
          }
        }
      },
...
      "ModelServerResponse": {
        "type": "object",
        "description": "Response for a single model server",
        "required": [
          "name",
          "kind",
          "extra_env",
          "user_declared"
        ],
        "properties": {
          "api_key_env": {
            "type": [
              "string",
              "null"
            ],
            "description": "Name of an env var providing the API key (e.g., `OLLAMA_API_KEY`)"
          },
          "base_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Base URL of the inference endpoint (e.g., `http://localhost:11434`)"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional display name for UI"
          },
          "extra_env": {
            "type": "object",
            "description": "Additional environment variables set when spawning agents that use this server",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "kind": {
            "type": "string",
            "description": "Kind: \"ollama\", \"openai-compat\", \"anthropic-api\", \"openai-api\", \"google-api\", \"lmstudio\""
          },
          "name": {
            "type": "string",
            "description": "Unique name (e.g., \"ollama-local\")"
          },
          "user_declared": {
            "type": "boolean",
            "description": "Whether this is a user-declared server (true) or an implicit builtin (false)"
          }
        }
      },
      "ModelServersResponse": {
        "type": "object",
        "description": "Response listing all model servers (declared + implicit builtins)",
        "required": [
          "servers",
          "total"
        ],
        "properties": {
          "servers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelServerResponse"
            },
            "description": "List of model servers"
          },
          "total": {
            "type": "integer",
            "description": "Total count",
            "minimum": 0
          }
        }
      },
...
      "StepResponse": {
        "type": "object",
        "description": "Response for a step",
        "required": [
          "name",
          "prompt",
          "outputs",
          "allowed_tools",
          "review_type",
          "permission_mode"
        ],
        "properties": {
          "allowed_tools": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "next_step": {
            "type": [
              "string",
              "null"
            ]
          },
          "outputs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permission_mode": {
            "type": "string"
          },
          "prompt": {
            "type": "string"
          },
          "review_type": {
            "type": "string",
            "description": "Type of review required: \"none\", \"plan\", \"visual\", \"pr\""
          }
        }
      },
...
      "UpdateIssueTypeRequest": {
        "type": "object",
        "description": "Request to update an issue type",
        "properties": {
          "color": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "fields": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/CreateFieldRequest"
            }
          },
          "glyph": {
            "type": [
              "string",
              "null"
            ]
          },
          "mode": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "project_required": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "steps": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/CreateStepRequest"
            }
          }
        }
      },
      "UpdateStepRequest": {
        "type": "object",
        "description": "Request to update a step",
        "properties": {
          "allowed_tools": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "next_step": {
            "type": [
              "string",
              "null"
            ]
          },
          "outputs": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "permission_mode": {
            "type": [
              "string",
              "null"
            ]
          },
          "prompt": {
            "type": [
              "string",
              "null"
            ]
          },
          "review_type": {
            "type": [
              "string",
              "null"
            ],
            "description": "Type of review required: \"none\", \"plan\", \"visual\", \"pr\""
          }
        }
      }
    }
  },
...
}